emacs: Get spacy
This is probably the 3rd or 4th time in half a year that I switch editor. No I
am not crazy!

When I switched to emacs I initially tried spacemacs, but the documentation was
not really useful for an absolute emacs beginner, so I went with doom emacs
which was a bit more straightforward.

But doom-emacs has weird performance issues for me, that even gccemacs cannot
really solve, so I tried out spacemacs. Turns out even with full config and lots
of layers its way faster than doom emacs.

So see you in two months when I find the next cool thing...

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>